pyproject.toml
[project]
name = "system-notification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "python-socketio>=5.16.0",
    "rethinkdb>=2.4.10.post1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
